requirements.txt
amqp==5.1.1
asgiref==3.6.0
async-timeout==4.0.2
backports.zoneinfo==0.2.1
billiard==3.6.4.0
Brotli==1.0.9
celery==5.2.7
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.6
cryptography==39.0.0
cssselect2==0.7.0
defusedxml==0.7.1
Deprecated==1.2.13
Django==4.1.4
django-ckeditor==6.5.1
django-crispy-forms==1.14.0
django-extensions==3.2.1
django-js-asset==2.0.0
django-localflavor==3.1
django-parler==2.3
django-rosetta==0.9.8
django-sendgrid==1.0.1
flower==1.1.0
fonttools==4.38.0
html5lib==1.1
humanize==4.4.0
idna==3.4
kombu==5.2.4
MarkupSafe==2.1.1
oauthlib==3.2.2
packaging==22.0
Pillow==9.2.0
polib==1.1.1
prometheus-client==0.15.0
prompt-toolkit==3.0.36
pycparser==2.21
pydyf==0.5.0
PyJWT==2.6.0
pyOpenSSL==23.0.0
pyphen==0.13.2
python-decouple==3.6
python-http-client==3.3.7
python-stdnum==1.18
python3-openid==3.2.0
pytz==2022.7
redis==4.3.4
requests==2.28.1
requests-oauthlib==1.3.1
sendgrid==6.9.7
six==1.16.0
social-auth-app-django==5.0.0
social-auth-core==4.3.0
sqlparse==0.4.3
starkbank-ecdsa==2.2.0
stripe==4.0.2
tinycss2==1.2.1
tornado==6.2
tzdata==2022.7
urllib3==1.26.13
vine==5.0.0
wcwidth==0.2.5
weasyprint==56.1
webencodings==0.5.1
Werkzeug==2.2.2
wrapt==1.14.1
zopfli==0.2.2
